Added a health script to the player character, as well as a cube that will damage the player.
Added a health script and a cube that will damage the player every 10 seconds. These showcase in the HealthScript and Exampledamagescript how to utilize the script for basic damage. There are no death functions or ui elements yet, this is simply barebones to allow for the creation and testing of the barrier powerup.